extract Exceptions into own package
