+++ test/imaginaryquadraticinteger/Exceptions/NonUniqueFactorizationDomainExceptionTest.java
@@ -14,8 +14,12 @@
  * You should have received a copy of the GNU General Public License along with 
  * this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package imaginaryquadraticinteger;
+package imaginaryquadraticinteger.Exceptions;
 
+import imaginaryquadraticinteger.Exceptions.NonUniqueFactorizationDomainException;
+import imaginaryquadraticinteger.ImaginaryQuadraticInteger;
+import imaginaryquadraticinteger.ImaginaryQuadraticRing;
+import imaginaryquadraticinteger.NumberTheoreticFunctionsCalculator;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.BeforeClass;
